gcp/requirements.txt
pillow == 8.0.0
tensorflow==2.18.0
google-cloud-storage==2.0.0
